ESQL: remove incorrect inline stats pruning (#141056)
Removes incorrect pruning in `inline stats` that lacked taking into
consideration the fact that groupings can be multi-valued. Fixes
https://github.com/elastic/elasticsearch/issues/139359#issuecomment-3773241566
(grouping by an mv fields can lead to inconsistencies in certain
queries) Fixes https://github.com/elastic/elasticsearch/issues/139359
Fixes https://github.com/elastic/elasticsearch/issues/140757

This also extracts a bunch of tests methods from the huge
`LogicalPlanOptimizerTests` and moves them to a new test class -
`PruneColumnsTests` which should include most of the tests that have in
the spotlight the `PruneColumns` optimization rule.

(cherry picked from commit 244538b2ce04ac186e93a71c31fa3f3654940646)
