+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/rules/logical/PushDownAndCombineFiltersTests.java
@@ -1373,17 +1373,8 @@ public class PushDownAndCombineFiltersTests extends AbstractLogicalPlanOptimizer
     /*
      * EsqlProject[[salary{f}#16, emp_no{f}#11]]
      * \_Limit[1000[INTEGER],false,false]
-     *   \_InlineJoin[LEFT,[salary{f}#16],[salary{r}#16]]
-     *     |_Filter[salary{f}#16 < 10000[INTEGER] AND salary{f}#16 > 10000[INTEGER]]
-     *     | \_EsRelation[employees][_meta_field{f}#17, emp_no{f}#11, first_name{f}#12, ..]
-     *     \_Aggregate[[salary{f}#16],[salary{f}#16]]
-     *       \_StubRelation[[_meta_field{f}#17, emp_no{f}#11, first_name{f}#12, gender{f}#13, hire_date{f}#18, job{f}#19, job.raw{f}#20, l
-     * anguages{f}#14, last_name{f}#15, long_noidx{f}#21, salary{f}#16]]
-     *
-     * stubReplacedSubPlan:
-     * Aggregate[[salary{f}#16],[salary{f}#16]]
-     * \_Filter[salary{f}#16 < 10000[INTEGER] AND salary{f}#16 > 10000[INTEGER]]
-     *   \_EsRelation[employees][_meta_field{f}#17, emp_no{f}#11, first_name{f}#12, ..]
+     *   \_Filter[salary{f}#16 < 10000[INTEGER] AND salary{f}#16 > 10000[INTEGER]]
+     *     \_EsRelation[employees][_meta_field{f}#17, emp_no{f}#11, first_name{f}#12, ..]
      */
     public void testPushDown_ImpossibleFilter_PastInlineJoin() {
         var plan = plan("""
@@ -1393,40 +1384,26 @@ public class PushDownAndCombineFiltersTests extends AbstractLogicalPlanOptimizer
             | WHERE salary > 10000
             | KEEP salary, emp_no
             """);
-        var subPlansResults = new HashSet<LocalRelation>();
-        var firstSubPlan = InlineJoin.firstSubPlan(plan, subPlansResults).stubReplacedSubPlan();
 
         var project = as(plan, EsqlProject.class);
         var limit = as(project.child(), Limit.class);
-        // InlineJoin
-        var ij = as(limit.child(), InlineJoin.class);
+        var filter = as(limit.child(), Filter.class);
+        assertThat(filter.condition(), instanceOf(And.class));
+        var and = as(filter.condition(), And.class);
+        var left = as(and.left(), LessThan.class);
+        var leftField = as(left.left(), FieldAttribute.class);
+        assertEquals("salary", leftField.name());
+        var leftRight = as(left.right(), Literal.class);
+        assertEquals(10000, leftRight.value());
+        var right = as(and.right(), GreaterThan.class);
+        var rightField = as(right.left(), FieldAttribute.class);
+        assertEquals("salary", rightField.name());
 
-        // InlineJoin left side
-        var left = as(ij.left(), Filter.class);
-        // Filter[salary{f}#16 < 10000[INTEGER] AND salary{f}#16 > 10000[INTEGER]]
-        assertThat(left.condition(), instanceOf(And.class));
-        var and = as(left.condition(), And.class);
-        var andLeft = as(and.left(), LessThan.class);
-        var andLeftField = as(andLeft.left(), FieldAttribute.class);
-        assertEquals("salary", andLeftField.name());
-        var andLeftRight = as(andLeft.right(), Literal.class);
-        assertEquals(10000, andLeftRight.value());
-        var andRight = as(and.right(), GreaterThan.class);
-        var andRightField = as(andRight.left(), FieldAttribute.class);
-        assertEquals("salary", andRightField.name());
-        var andRightRight = as(andRight.right(), Literal.class);
-        assertEquals(10000, andRightRight.value());
-        as(left.child(), EsRelation.class);
+        assertEquals(leftField, rightField);
 
-        // InlineJoin right side
-        var right = as(ij.right(), Aggregate.class);
-
-        // What EsqlSession is doing
-        var firstSubPlanAggregate = as(firstSubPlan, Aggregate.class);
-        assertEquals(right.output(), firstSubPlanAggregate.output());
-        var firstSubPlanFilter = as(firstSubPlanAggregate.child(), Filter.class);
-        // important bit below: the filter that is executed in the right hand side is the same as the one in the left hand side
-        assertEquals(left, firstSubPlanFilter);
+        var rightRight = as(right.right(), Literal.class);
+        assertEquals(10000, rightRight.value());
+        as(filter.child(), EsRelation.class);
 
         assertWarnings("Line 3:16: Field 'salary' shadowed by field at line 3:40", "No limit defined, adding default limit of [1000]");
     }
